Estimate maximum lambda values numerically. First starts with a large initial value max_lambda_mean and max_lambda_prob, and runs the EM algorithm on decreasing set of values (sequentially halved). This stops once you see non-simple probabilities or means, and returns the smallest regularization (lambda) value pair that gives full sparsity.

Note that the zero_stabilize=TRUE option is used in flowtrend(), which basically means the EM algorithm runs only until the zero pattern stabilizes.

Usage

calc_max_lambda(
  ylist,
  countslist = NULL,
  numclust,
  max_lambda_mean = 4000,
  max_lambda_prob = 1000,
  verbose = FALSE,
  iimax = 16,
  ...
)

Arguments

ylist

List of responses.

numclust

Number of clusters.

max_lambda_mean

Defaults to 4000.

max_lambda_prob

Defaults to 1000.

iimax

Maximum value of x for 2^-x factors to try.

...

Other arguments to flowtrend_once().

Value

list containing the two maximum values to use.