Cross-validation for flowtrend(). Saves results to separate files in destin.

Cross-validation for flowtrend(). Saves results to separate files in destin.

Usage

cv_flowtrend(
  ylist,
  countslist,
  x = NULL,
  destin = ".",
  lambda_means,
  lambda_probs,
  l,
  l_prob,
  iimat = NULL,
  maxdev,
  numclust,
  nfold,
  blocksize,
  nrestart,
  verbose = FALSE,
  refit = FALSE,
  save_meta = FALSE,
  mc.cores = 1,
  folds = NULL,
  seedtab = NULL,
  niter = 1000,
  ...
)

Arguments

ylist

Data.

countslist

Counts corresponding to multiplicities.

x

Times, if points are not evenly spaced. Defaults to NULL, in which case the value becomes 1:T, for the $T==length(ylist)$.

destin

Directory where output files are saved.

lambda_means

Regularization parameters for means.

lambda_probs

Regularization parameters for probs.

l

Degree of differencing for the mean trend filtering. l=0 will give you piecewise constant means; l=1 is piecewise linear, and so forth.

l_prob

Degree of differencing for the probability trend filtering.

maxdev

Maximum deviation of cluster means across time..

numclust

Number of clusters.

nfold

Number of cross-validation folds. Defaults to 5.

blocksize

Contiguous time blocks from which to form CV time folds.

nrestart

Number of repetitions.

verbose

Loud or not? EM iteration progress is printed.

refit

If TRUE, estimate the model on the full data, for each pair of regularization parameters.

save_meta

If TRUE, save meta data.

mc.cores

Use this many CPU cores.

folds

Manually provide CV folds (list of time points of data to use as CV folds). Defaults to NULL.

niter

Maximum number of EM iterations.

...

Additional arguments to flowtrend().

Value

No return.