Main function for summarizing the cross-validation results.

Main function for summarizing the cross-validation results.

Usage

cv_summary(destin = ".", save = FALSE, filename = "summary.RDS")

Arguments

destin

Directory where output files are saved.

save

If TRUE, save to file.path(destin, filename).

filename

File name to save to.

Value

List containing summarized results from cross-validation. Here are some objects in this list: bestres is the the overall best model chosen from the cross-validation; cvscoremat is a 2d matrix of CV scores from all pairs of regularization parameters; bestreslist is a list of all the best models (out of nrestart EM replications) from the each pair of lambda values. If isTRUE(save), nothing is returned.