Estimate flowtrend model once.

Estimate flowtrend model once.

Usage

flowtrend_once(
  ylist,
  countslist = NULL,
  x = NULL,
  numclust,
  niter = 1000,
  l,
  l_prob = NULL,
  mn = NULL,
  lambda = 0,
  lambda_prob = NULL,
  verbose = FALSE,
  tol_em = 1e-04,
  maxdev = NULL,
  countslist_overwrite = NULL,
  admm = TRUE,
  admm_err_rel = 0.001,
  admm_err_abs = 1e-04,
  admm_local_adapt = TRUE,
  admm_local_adapt_niter = if (admm_local_adapt) 10 else 1,
  rho_init = 0.1,
  check_convergence = TRUE,
  seed = NULL
)

Arguments

ylist

Data.

countslist

Counts corresponding to multiplicities.

x

Times, if points are not evenly spaced. Defaults to NULL, in which case the value becomes 1:T, for the $T==length(ylist)$.

numclust

Number of clusters.

niter

Maximum number of EM iterations.

l

Degree of differencing for the mean trend filtering. l=0 will give you piecewise constant means; l=1 is piecewise linear, and so forth.

l_prob

Degree of differencing for the probability trend filtering.

mn

Initial value for cluster means. Defaults to NULL, in which case initial values are randomly chosen from the data.

lambda

Smoothing parameter for means

lambda_prob

Smoothing parameter for probabilities

verbose

Loud or not? EM iteration progress is printed.

tol_em

Relative numerical improvement of the objective value at which to stop the EM algorithm

maxdev

Maximum deviation of cluster means across time..

admm_local_adapt_niter

Value

List object with flowtrend model estimates.