A helper function to print the progress of a loop or simulation.

A helper function to print the progress of a loop or simulation.

Usage

print_progress(
  isim,
  nsim,
  type = "simulation",
  lapsetime = NULL,
  lapsetimeunit = "seconds",
  start.time = NULL,
  fill = FALSE
)

Arguments

isim

Replicate number.

nsim

Total number of replicates.

type

Type of job you're running. Defaults to "simulation".

lapsetime

Lapsed time, in seconds (by default).

lapsetimeunit

"second".

start.time

start time.

fill

Whether or not to fill the line.

Value

No return